+++ src/rpc/client/index.ts
@@ -1,5 +1,4 @@
 /** @module RPC/Client */
 
 export * from './client-management.js';
-export * from './local-fallback-client.js';
 export * from './types.js';